+++ src/functions.py
@@ -1,33 +1,6 @@
 # --------------------------------------------------------------------
-# Function and classes called by scenes
+# Helper functions
 # --------------------------------------------------------------------
-
-#
-# This class remove duplicate midi message by taking care of an offset logic
-# NOT STABLE SUSPECT OVERFLOW 
-class RemoveDuplicates:
-    def __init__(self, _wait=0):
-        self.wait = _wait
-        self.prev_ev = None
-        self.prev_time = 0
-
-    def __call__(self, ev):
-        if ev.type == NOTEOFF:
-            sleep(self.wait)
-            return ev
-        now = engine.time()
-        offset = now - self.prev_time
-        if offset >= 0.035:
-            # if ev.type == NOTEON:
-            #    print "+ " + str(offset)
-            r = ev
-        else:
-            # if ev.type == NOTEON:
-            #    print "- " + str(offset)
-            r = None
-        self.prev_ev = ev
-        self.prev_time = now
-        return r
 
 # -------------------------------------------------------------------------------------------
 '''
@@ -52,13 +25,16 @@ def glissando(ev, from_note, to_note, vel, duration, direction, port):
 
 # -------------------------------------------------------------------------------------------
 
-
-# Navigate through secenes and subscenes
 def NavigateToScene(ev):
-    # MIDIDINGS does not wrap in the builtin ScenesSwitch but SubSecenesSwitch yes with the wrap parameter
-    # With that function, you can wrap trough Scenes AND SubScenes
-    # That function assume that the first SceneNumber is 1
-    # TODO field, values = dict(scenes()).items()[0]
+    ''' 
+    Navigate through Scenes and Sub-Scenes
+    
+    MIDIDINGS does not wrap in the builtin ScenesSwitch but SubSecenesSwitch yes with the wrap parameter
+    
+    With that function, you can wrap trough Scenes AND SubScenes
+    
+    That function assume that the first SceneNumber is 1
+    '''
     if ev.ctrl == 20:
         nb_scenes = len(scenes())
         cs = current_scene()
@@ -85,24 +61,6 @@ def NavigateToScene(ev):
                 switch_subscene(css + 1)
             else:
                 switch_subscene(1)
-
-
-# Stop any audio processing, managed by a simple bash script
-def AllAudioOff(ev):
-    return "/bin/bash ./kill.sh"
-
-
-# Audio and midi players suitable for my SD-90
-def play_file(filename):
-    fname, fext = os.path.splitext(filename)
-    if fext == ".mp3":
-        path = " /tmp/soundlib/mp3/"
-        command = "mpg123 -q"
-    elif fext == ".mid":
-        path = " /tmp/soundlib/midi/"
-        command = "aplaymidi -p 20:1"
-
-    return command + path + filename
 
 
 # Create a pitchbend from a filter logic
