+++ src/main.py
@@ -40,7 +40,7 @@ net_config=plugins['net']
 config(
 
     # Defaults
-    # initial_scene = 2,
+    # initial_scene = 1,
     # backend = 'alsa',
     # client_name = 'mididings',
 
@@ -48,7 +48,7 @@ config(
     #   Device name                     # Description               #
     #  
 
-    # Ports are tokenized and replaced by script_builder.sh
+    # Ports are tokenized and sed/awk by script_builder.sh
 
     out_ports = [
