+++ src/proto/remove_duplicates.py
@@ -1,0 +1,25 @@
+# This class remove duplicate midi message by taking care of an offset logic
+# NOT STABLE SUSPECT OVERFLOW 
+class RemoveDuplicates:
+    def __init__(self, _wait=0):
+        self.wait = _wait
+        self.prev_ev = None
+        self.prev_time = 0
+
+    def __call__(self, ev):
+        if ev.type == NOTEOFF:
+            sleep(self.wait)
+            return ev
+        now = engine.time()
+        offset = now - self.prev_time
+        if offset >= 0.035:
+            # if ev.type == NOTEON:
+            #    print "+ " + str(offset)
+            r = ev
+        else:
+            # if ev.type == NOTEON:
+            #    print "- " + str(offset)
+            r = None
+        self.prev_ev = ev
+        self.prev_time = now
+        return r
